most of the way to a rewire
