+++ spynnaker/pyNN/models/spike_source/spike_source_poisson_vertex.py
@@ -27,27 +27,22 @@ from pacman.model.resources import (
     ConstantSDRAM, CPUCyclesPerTickResource, DTCMResource, ResourceContainer)
 from spinn_front_end_common.abstract_models import (
     AbstractChangableAfterRun, AbstractProvidesOutgoingPartitionConstraints,
-    AbstractGeneratesDataSpecification,
     AbstractRewritesDataSpecification)
 from spinn_front_end_common.abstract_models.impl import (
     ProvidesKeyToAtomMappingImpl, TDMAAwareApplicationVertex)
-from spinn_front_end_common.interface.simulation import simulation_utilities
 from spinn_front_end_common.interface.buffer_management import (
     recording_utilities)
 from spinn_front_end_common.utilities import (
     helpful_functions, globals_variables)
 from spinn_front_end_common.utilities.constants import (
-    SYSTEM_BYTES_REQUIREMENT, SIMULATION_N_BYTES, BYTES_PER_WORD,
-    MICRO_TO_MILLISECOND_CONVERSION, MICRO_TO_SECOND_CONVERSION)
-from spinn_front_end_common.utilities.exceptions import ConfigurationException
+    SYSTEM_BYTES_REQUIREMENT, MICRO_TO_SECOND_CONVERSION)
 from spinn_front_end_common.interface.profiling import profile_utils
 from spynnaker.pyNN.models.common import (
     AbstractSpikeRecordable, MultiSpikeRecorder, SimplePopulationSettable)
-from spynnaker.pyNN.utilities import constants
 from spynnaker.pyNN.models.abstract_models import (
     AbstractReadParametersBeforeSet)
 from .spike_source_poisson_machine_vertex import (
-    SpikeSourcePoissonMachineVertex)
+    SpikeSourcePoissonMachineVertex, _flatten, get_rates_bytes)
 from spynnaker.pyNN.utilities.utility_calls import create_mars_kiss_seeds
 from spynnaker.pyNN.utilities.struct import Struct
 from spynnaker.pyNN.utilities.ranged.spynnaker_ranged_dict \
@@ -56,13 +51,6 @@ from spynnaker.pyNN.utilities.ranged.spynnaker_ranged_list \
     import SpynnakerRangedList
 
 logger = logging.getLogger(__name__)
-
-# 1. uint32_t has_key; 2. uint32_t key; 3. uint32_t set_rate_neuron_id_mask;
-# 4. UFRACT seconds_per_tick; 5. REAL ticks_per_second;
-# 6. REAL slow_rate_per_tick_cutoff; 7. REAL fast_rate_per_tick_cutoff;
-# 8. uint32_t first_source_id; 9. uint32_t n_spike_sources;
-# 10,11,12,13 mars_kiss64_seed_t (uint[4]) spike_source_seed;
-PARAMS_BASE_WORDS = 13
 
 # Seed offset in parameters and size on bytes
 SEED_OFFSET_BYTES = 9 * 4
@@ -75,16 +63,13 @@ PARAMS_WORDS_PER_NEURON = 2
 # sqrt_lambda, isi_val, time_to_spike
 PARAMS_WORDS_PER_RATE = 8
 
-SLOW_RATE_PER_TICK_CUTOFF = 0.01  # as suggested by MH (between Exp and Knuth)
-FAST_RATE_PER_TICK_CUTOFF = 10  # between Knuth algorithm and Gaussian approx.
 _REGIONS = SpikeSourcePoissonMachineVertex.POISSON_SPIKE_SOURCE_REGIONS
+SLOW_RATE_PER_TICK_CUTOFF = (
+    SpikeSourcePoissonMachineVertex.SLOW_RATE_PER_TICK_CUTOFF)
 OVERFLOW_TIMESTEPS_FOR_SDRAM = 5
 
 # The microseconds per timestep will be divided by this to get the max offset
 _MAX_OFFSET_DENOMINATOR = 10
-
-# The maximum timestep - this is the maximum value of a uint32
-_MAX_TIMESTEP = 0xFFFFFFFF
 
 
 _PoissonStruct = Struct([
@@ -98,28 +83,16 @@ _PoissonStruct = Struct([
     DataType.UINT32])  # timesteps to next spike
 
 
-def _flatten(alist):
-    for item in alist:
-        if hasattr(item, "__iter__"):
-            for subitem in _flatten(item):
-                yield subitem
-        else:
-            yield item
-
-
 class SpikeSourcePoissonVertex(
-        TDMAAwareApplicationVertex, AbstractGeneratesDataSpecification,
-        AbstractSpikeRecordable, AbstractProvidesOutgoingPartitionConstraints,
+        TDMAAwareApplicationVertex, AbstractSpikeRecordable,
+        AbstractProvidesOutgoingPartitionConstraints,
         AbstractChangableAfterRun, AbstractReadParametersBeforeSet,
-        AbstractRewritesDataSpecification, SimplePopulationSettable,
-        ProvidesKeyToAtomMappingImpl):
+        SimplePopulationSettable, ProvidesKeyToAtomMappingImpl):
     """ A Poisson Spike source object
     """
     __slots__ = [
         "__change_requires_mapping",
-        "__change_requires_neuron_parameters_reload",
         "__duration",
-        "__machine_time_step",
         "__model",
         "__model_name",
         "__n_atoms",
@@ -173,7 +146,6 @@ class SpikeSourcePoissonVertex(
 
         # check for changes parameters
         self.__change_requires_mapping = True
-        self.__change_requires_neuron_parameters_reload = False
 
         self.__spike_recorder = MultiSpikeRecorder()
 
@@ -293,7 +265,6 @@ class SpikeSourcePoissonVertex(
             use_list_as_value=not hasattr(time_to_spike[0], "__len__"))
         self.__rng = numpy.random.RandomState(seed)
         self.__rate_change = numpy.zeros(n_neurons)
-        self.__machine_time_step = None
 
         # get config from simulator
         config = globals_variables.get_simulator().config
@@ -318,6 +289,10 @@ class SpikeSourcePoissonVertex(
                 [numpy.max(r) for r in self.__data["rates"]])
             self.__max_spikes = numpy.sum(scipy.stats.poisson.ppf(
                 1.0 - (1.0 / max_rates), max_rates))
+
+    @property
+    def n_profile_samples(self):
+        return self.__n_profile_samples
 
     @property
     def rate(self):
@@ -424,6 +399,14 @@ class SpikeSourcePoissonVertex(
         raise Exception("Set the duration of a Poisson source using duration")
 
     @property
+    def time_to_spike(self):
+        return self.__data["time_to_spike"]
+
+    @property
+    def rate_change(self):
+        return self.__rate_change
+
+    @property
     @overrides(AbstractChangableAfterRun.requires_mapping)
     def requires_mapping(self):
         return self.__change_requires_mapping
@@ -435,7 +418,9 @@ class SpikeSourcePoissonVertex(
     @overrides(SimplePopulationSettable.set_value)
     def set_value(self, key, value):
         SimplePopulationSettable.set_value(self, key, value)
-        self.__change_requires_neuron_parameters_reload = True
+        for machine_vertex in self.machine_vertices:
+            if isinstance(machine_vertex, AbstractRewritesDataSpecification):
+                machine_vertex.set_reload_required(True)
 
     def max_spikes_per_ts(self, machine_time_step):
         """
@@ -477,7 +462,7 @@ class SpikeSourcePoissonVertex(
         :param int machine_time_step:
         """
         # pylint: disable=arguments-differ
-        poisson_params_sz = self.get_rates_bytes(vertex_slice)
+        poisson_params_sz = get_rates_bytes(vertex_slice, self.__data["rates"])
         other = ConstantSDRAM(
             SYSTEM_BYTES_REQUIREMENT +
             SpikeSourcePoissonMachineVertex.get_provenance_data_size(0) +
@@ -525,264 +510,6 @@ class SpikeSourcePoissonVertex(
         self.__kiss_seed = dict()
         self.__rng = None
 
-    def get_rates_bytes(self, vertex_slice):
-        """ Gets the size of the Poisson rates in bytes
-
-        :param ~pacman.model.graphs.common.Slice vertex_slice:
-        :rtype: int
-        """
-        n_rates = sum(len(self.__data["rates"][i]) for i in range(
-            vertex_slice.lo_atom, vertex_slice.hi_atom + 1))
-        return ((vertex_slice.n_atoms * PARAMS_WORDS_PER_NEURON) +
-                (n_rates * PARAMS_WORDS_PER_RATE)) * BYTES_PER_WORD
-
-    def reserve_memory_regions(self, spec, placement):
-        """ Reserve memory regions for Poisson source parameters and output\
-            buffer.
-
-        :param ~data_specification.DataSpecification spec:
-            the data specification writer
-        :param ~pacman.models.placements.Placement placement:
-            the location this vertex resides on in the machine
-        :return: None
-        """
-        spec.comment("\nReserving memory space for data regions:\n\n")
-
-        # Reserve memory:
-        spec.reserve_memory_region(
-            region=_REGIONS.SYSTEM_REGION.value,
-            size=SIMULATION_N_BYTES,
-            label='setup')
-
-        # reserve poisson parameters and rates DSG region
-        self._reserve_poisson_params_rates_region(placement, spec)
-
-        spec.reserve_memory_region(
-            region=_REGIONS.SPIKE_HISTORY_REGION.value,
-            size=recording_utilities.get_recording_header_size(1),
-            label="Recording")
-
-        profile_utils.reserve_profile_region(
-            spec, _REGIONS.PROFILER_REGION.value, self.__n_profile_samples)
-
-        spec.reserve_memory_region(
-            region=_REGIONS.TDMA_REGION.value, label="tdma_region",
-            size=self.tdma_sdram_size_in_bytes)
-
-        placement.vertex.reserve_provenance_data_region(spec)
-
-    def _reserve_poisson_params_rates_region(self, placement, spec):
-        """ Allocate space for the Poisson parameters and rates regions as\
-            they can be reused for setters after an initial run
-
-        :param ~pacman.models.placements.Placement placement:
-            the location on machine for this vertex
-        :param ~data_specification.DataSpecification spec: the DSG writer
-        :return: None
-        """
-        spec.reserve_memory_region(
-            region=_REGIONS.POISSON_PARAMS_REGION.value,
-            size=PARAMS_BASE_WORDS * 4, label="PoissonParams")
-        spec.reserve_memory_region(
-            region=_REGIONS.RATES_REGION.value,
-            size=self.get_rates_bytes(placement.vertex.vertex_slice),
-            label='PoissonRates')
-
-    def _write_poisson_parameters(
-            self, spec, graph, placement, routing_info,
-            vertex_slice, machine_time_step):
-        """ Generate Parameter data for Poisson spike sources
-
-        :param ~data_specification.DataSpecification spec:
-            the data specification writer
-        :param ~pacman.model.graphs.machine.MachineGraph graph:
-        :param ~pacman.model.placements.Placement placement:
-        :param ~pacman.model.routing_info.RoutingInfo routing_info:
-        :param ~pacman.model.graphs.common.Slice vertex_slice:
-            the slice of atoms a machine vertex holds from its application
-            vertex
-        :param int machine_time_step: the time between timer tick updates.
-        """
-        # pylint: disable=too-many-arguments, too-many-locals
-        spec.comment("\nWriting Parameters for {} poisson sources:\n"
-                     .format(vertex_slice.n_atoms))
-
-        # Set the focus to the memory region 2 (neuron parameters):
-        spec.switch_write_focus(_REGIONS.POISSON_PARAMS_REGION.value)
-
-        # Write Key info for this core:
-        key = routing_info.get_first_key_from_pre_vertex(
-            placement.vertex, constants.SPIKE_PARTITION_ID)
-        spec.write_value(data=1 if key is not None else 0)
-        spec.write_value(data=key if key is not None else 0)
-
-        # Write the incoming mask if there is one
-        in_edges = graph.get_edges_ending_at_vertex_with_partition_name(
-            placement.vertex, constants.LIVE_POISSON_CONTROL_PARTITION_ID)
-        if len(in_edges) > 1:
-            raise ConfigurationException(
-                "Only one control edge can end at a Poisson vertex")
-        incoming_mask = 0
-        if len(in_edges) == 1:
-            in_edge = in_edges[0]
-
-            # Get the mask of the incoming keys
-            incoming_mask = \
-                routing_info.get_routing_info_for_edge(in_edge).first_mask
-            incoming_mask = ~incoming_mask & 0xFFFFFFFF
-        spec.write_value(incoming_mask)
-
-        # Write the number of seconds per timestep (unsigned long fract)
-        spec.write_value(
-            data=float(machine_time_step) / MICRO_TO_SECOND_CONVERSION,
-            data_type=DataType.U032)
-
-        # Write the number of timesteps per second (integer)
-        spec.write_value(
-            data=int(MICRO_TO_SECOND_CONVERSION / float(machine_time_step)))
-
-        # Write the slow-rate-per-tick-cutoff (accum)
-        spec.write_value(
-            data=SLOW_RATE_PER_TICK_CUTOFF, data_type=DataType.S1615)
-
-        # Write the fast-rate-per-tick-cutoff (accum)
-        spec.write_value(
-            data=FAST_RATE_PER_TICK_CUTOFF, data_type=DataType.S1615)
-
-        # Write the lo_atom ID
-        spec.write_value(data=vertex_slice.lo_atom)
-
-        # Write the number of sources
-        spec.write_value(data=vertex_slice.n_atoms)
-
-        # Write the random seed (4 words), generated randomly!
-        if vertex_slice not in self.__kiss_seed:
-            self.__kiss_seed[vertex_slice] = create_mars_kiss_seeds(
-                self.__rng, self.__seed)
-        for value in self.__kiss_seed[vertex_slice]:
-            spec.write_value(data=value)
-
-    def _write_poisson_rates(self, spec, vertex_slice, machine_time_step,
-                             first_machine_time_step):
-        """ Generate Rate data for Poisson spike sources
-
-        :param ~data_specification.DataSpecification spec:
-            the data specification writer
-        :param ~pacman.model.graphs.common.Slice vertex_slice:
-            the slice of atoms a machine vertex holds from its application
-            vertex
-        :param int machine_time_step: the time between timer tick updates.
-        :param int first_machine_time_step:
-            First machine time step to start from the correct index
-        """
-        spec.comment("\nWriting Rates for {} poisson sources:\n"
-                     .format(vertex_slice.n_atoms))
-
-        # Set the focus to the memory region 2 (neuron parameters):
-        spec.switch_write_focus(_REGIONS.RATES_REGION.value)
-
-        # Extract the data on which to work and convert to appropriate form
-        starts = numpy.array(list(_flatten(
-            self.__data["starts"][vertex_slice.as_slice]))).astype("float")
-        durations = numpy.array(list(_flatten(
-            self.__data["durations"][vertex_slice.as_slice]))).astype("float")
-        local_rates = self.__data["rates"][vertex_slice.as_slice]
-        n_rates = numpy.array([len(r) for r in local_rates])
-        splits = numpy.cumsum(n_rates)
-        rates = numpy.array(list(_flatten(local_rates)))
-        time_to_spike = numpy.array(list(_flatten(
-            self.__data["time_to_spike"][vertex_slice.as_slice]))).astype("u4")
-        rate_change = self.__rate_change[vertex_slice.as_slice]
-
-        # Convert start times to start time steps
-        starts_scaled = self._convert_ms_to_n_timesteps(
-            starts, machine_time_step)
-
-        # Convert durations to end time steps, using the maximum for "None"
-        # duration (which means "until the end")
-        no_duration = numpy.isnan(durations)
-        durations_filtered = numpy.where(no_duration, 0, durations)
-        ends_scaled = self._convert_ms_to_n_timesteps(
-            durations_filtered, machine_time_step) + starts_scaled
-        ends_scaled = numpy.where(no_duration, _MAX_TIMESTEP, ends_scaled)
-
-        # Work out the timestep at which the next rate activates, using
-        # the maximum value at the end (meaning there is no "next")
-        starts_split = numpy.array_split(starts_scaled, splits)
-        next_scaled = numpy.concatenate([numpy.append(s[1:], _MAX_TIMESTEP)
-                                         for s in starts_split[:-1]])
-
-        # Compute the spikes per tick for each rate for each atom
-        spikes_per_tick = rates * (float(machine_time_step) /
-                                   MICRO_TO_SECOND_CONVERSION)
-
-        # Determine the properties of the sources
-        is_fast_source = spikes_per_tick >= SLOW_RATE_PER_TICK_CUTOFF
-        is_faster_source = spikes_per_tick >= FAST_RATE_PER_TICK_CUTOFF
-        not_zero = spikes_per_tick > 0
-        # pylint: disable=assignment-from-no-return
-        is_slow_source = numpy.logical_not(is_fast_source)
-
-        # Compute the e^-(spikes_per_tick) for fast sources to allow fast
-        # computation of the Poisson distribution to get the number of
-        # spikes per timestep
-        exp_minus_lambda = DataType.U032.encode_as_numpy_int_array(
-            numpy.where(is_fast_source, numpy.exp(-1.0 * spikes_per_tick), 0))
-
-        # Compute sqrt(lambda) for "faster" sources to allow Gaussian
-        # approximation of the Poisson distribution to get the number of
-        # spikes per timestep
-        sqrt_lambda = DataType.S1615.encode_as_numpy_int_array(
-            numpy.where(is_faster_source, numpy.sqrt(spikes_per_tick), 0))
-
-        # Compute the inter-spike-interval for slow sources to get the
-        # average number of timesteps between spikes
-        isi_val = numpy.where(
-            not_zero & is_slow_source,
-            (1.0 / spikes_per_tick).astype(int), 0).astype("uint32")
-
-        # Reuse the time-to-spike read from the machine (if has been run)
-        # or don't if the rate has since been changed
-        time_to_spike_split = numpy.array_split(time_to_spike, splits)
-        time_to_spike = numpy.concatenate(
-            [t if rate_change[i] else numpy.repeat(0, len(t))
-             for i, t in enumerate(time_to_spike_split[:-1])])
-
-        # Turn the fast source booleans into uint32
-        is_fast_source = is_fast_source.astype("uint32")
-
-        # Group together the rate data for the core by rate
-        core_data = numpy.dstack((
-            starts_scaled, ends_scaled, next_scaled, is_fast_source,
-            exp_minus_lambda, sqrt_lambda, isi_val, time_to_spike))[0]
-
-        # Group data by neuron id
-        core_data_split = numpy.array_split(core_data, splits)
-
-        # Work out the index where the core should start based on the given
-        # first timestep
-        ends_scaled_split = numpy.array_split(ends_scaled, splits)
-        indices = [numpy.argmax(e > first_machine_time_step)
-                   for e in ends_scaled_split[:-1]]
-
-        # Build the final data for this core, and write it
-        final_data = numpy.concatenate([
-            numpy.concatenate(([len(d), indices[i]], numpy.concatenate(d)))
-            for i, d in enumerate(core_data_split[:-1])])
-        spec.write_array(final_data)
-
-    @staticmethod
-    def _convert_ms_to_n_timesteps(value, machine_time_step):
-        return numpy.round(
-            value * (MICRO_TO_MILLISECOND_CONVERSION /
-                     float(machine_time_step))).astype("uint32")
-
-    @staticmethod
-    def _convert_n_timesteps_to_ms(value, machine_time_step):
-        return (
-            value / (MICRO_TO_MILLISECOND_CONVERSION /
-                     float(machine_time_step)))
-
     @overrides(AbstractSpikeRecordable.is_recording_spikes)
     def is_recording_spikes(self):
         return self.__spike_recorder.record
@@ -812,60 +539,18 @@ class SpikeSourcePoissonVertex(
     def get_cpu_usage_for_atoms():
         return 0
 
-    @inject_items({
-        "machine_time_step": "MachineTimeStep",
-        "routing_info": "MemoryRoutingInfos",
-        "graph": "MemoryMachineGraph",
-        "first_machine_time_step": "FirstMachineTimeStep"})
+    def kiss_seed(self, vertex_slice):
+        if vertex_slice not in self.__kiss_seed:
+            self.__kiss_seed[vertex_slice] = create_mars_kiss_seeds(
+                self.__rng, self.__seed)
+        return self.__kiss_seed[vertex_slice]
+
+    @inject_items({"machine_time_step": "MachineTimeStep"})
     @overrides(
-        AbstractRewritesDataSpecification.regenerate_data_specification,
-        additional_arguments={
-            "machine_time_step", "routing_info", "graph",
-            "first_machine_time_step"})
-    def regenerate_data_specification(
-            self, spec, placement, machine_time_step, routing_info, graph,
-            first_machine_time_step):
-        """
-        :param int machine_time_step:
-        :param ~pacman.model.routing_info.RoutingInfo routing_info:
-        :param ~pacman.model.graphs.machine.MachineGraph graph:
-        :param int first_machine_time_step:
-        """
-        # pylint: disable=too-many-arguments, arguments-differ
-
-        # reserve the neuron parameters data region
-        self._reserve_poisson_params_rates_region(placement, spec)
-
-        # write parameters
-        self._write_poisson_parameters(
-            spec=spec, graph=graph, placement=placement,
-            routing_info=routing_info,
-            vertex_slice=placement.vertex.vertex_slice,
-            machine_time_step=machine_time_step)
-
-        # write rates
-        self._write_poisson_rates(
-            spec, placement.vertex.vertex_slice, machine_time_step,
-            first_machine_time_step)
-
-        # end spec
-        spec.end_specification()
-
-    @inject_items({"first_machine_time_step": "FirstMachineTimeStep"})
-    @overrides(AbstractRewritesDataSpecification.reload_required,
-               additional_arguments={"first_machine_time_step"})
-    def reload_required(self, first_machine_time_step):
-        # pylint: disable=arguments-differ
-        return (self.__change_requires_neuron_parameters_reload or
-                first_machine_time_step == 0)
-
-    @overrides(AbstractRewritesDataSpecification.reload_required)
-    def set_reload_required(self, new_value):
-        self.__change_requires_neuron_parameters_reload = new_value
-
-    @overrides(AbstractReadParametersBeforeSet.read_parameters_from_machine)
+        AbstractReadParametersBeforeSet.read_parameters_from_machine,
+        additional_arguments=["machine_time_step"])
     def read_parameters_from_machine(
-            self, transceiver, placement, vertex_slice):
+            self, transceiver, placement, vertex_slice, machine_time_step):
 
         # locate SDRAM address where parameters are stored
         poisson_params = \
@@ -882,7 +567,8 @@ class SpikeSourcePoissonVertex(
                 placement, _REGIONS.RATES_REGION.value, transceiver)
 
         # get size of poisson params
-        size_of_region = self.get_rates_bytes(vertex_slice)
+        size_of_region = self.get_rates_bytes(
+            vertex_slice, self.__data["rates"])
 
         # get data from the machine
         byte_array = transceiver.read_memory(
@@ -920,82 +606,12 @@ class SpikeSourcePoissonVertex(
             self.__data["rates"].set_value_by_id(
                 i,
                 spikes_per_tick *
-                (MICRO_TO_SECOND_CONVERSION / float(self.__machine_time_step)))
+                (MICRO_TO_SECOND_CONVERSION / float(machine_time_step)))
 
             # Store the updated time until next spike so that it can be
             # rewritten when the parameters are loaded
             self.__data["time_to_spike"].set_value_by_id(
                 i, time_to_next_spike)
-
-    @inject_items({
-        "machine_time_step": "MachineTimeStep",
-        "time_scale_factor": "TimeScaleFactor",
-        "routing_info": "MemoryRoutingInfos",
-        "data_n_time_steps": "DataNTimeSteps",
-        "graph": "MemoryMachineGraph",
-        "first_machine_time_step": "FirstMachineTimeStep"
-    })
-    @overrides(
-        AbstractGeneratesDataSpecification.generate_data_specification,
-        additional_arguments={
-            "machine_time_step", "time_scale_factor", "routing_info",
-            "data_n_time_steps", "graph", "first_machine_time_step"
-        }
-    )
-    def generate_data_specification(
-            self, spec, placement, machine_time_step, time_scale_factor,
-            routing_info, data_n_time_steps, graph, first_machine_time_step):
-        """
-        :param int machine_time_step:
-        :param int time_scale_factor:
-        :param ~pacman.model.routing_info.RoutingInfo routing_info:
-        :param int data_n_time_steps:
-        :param ~pacman.model.graphs.machine.MachineGraph graph:
-        :param int first_machine_time_step:
-        """
-        # pylint: disable=too-many-arguments, arguments-differ
-        self.__machine_time_step = machine_time_step
-        vertex_slice = placement.vertex.vertex_slice
-
-        spec.comment("\n*** Spec for SpikeSourcePoisson Instance ***\n\n")
-
-        # Reserve SDRAM space for memory areas:
-        self.reserve_memory_regions(spec, placement)
-
-        # write setup data
-        spec.switch_write_focus(_REGIONS.SYSTEM_REGION.value)
-        spec.write_array(simulation_utilities.get_simulation_header_array(
-            placement.vertex.get_binary_file_name(), machine_time_step,
-            time_scale_factor))
-
-        # write recording data
-        spec.switch_write_focus(_REGIONS.SPIKE_HISTORY_REGION.value)
-        sdram = self.get_recording_sdram_usage(vertex_slice, machine_time_step)
-        recorded_region_sizes = [sdram.get_total_sdram(data_n_time_steps)]
-        spec.write_array(recording_utilities.get_recording_header_array(
-            recorded_region_sizes))
-
-        # write parameters
-        self._write_poisson_parameters(
-            spec, graph, placement, routing_info, vertex_slice,
-            machine_time_step)
-
-        # write rates
-        self._write_poisson_rates(spec, vertex_slice, machine_time_step,
-                                  first_machine_time_step)
-
-        # write profile data
-        profile_utils.write_profile_region_data(
-            spec, _REGIONS.PROFILER_REGION.value,
-            self.__n_profile_samples)
-
-        # write tdma params
-        spec.switch_write_focus(_REGIONS.TDMA_REGION.value)
-        spec.write_array(self.generate_tdma_data_specification_data(
-            self.vertex_slices.index(vertex_slice)))
-
-        # End-of-Spec:
-        spec.end_specification()
 
     @overrides(AbstractSpikeRecordable.get_spikes)
     def get_spikes(self, placements, buffer_manager, machine_time_step):
